#5215: Get rid of `FillMatchArmsIntention`
Empty match expressions filling is now implemented
in `RsNonExhaustiveMatchInspection`
