--- src/main/resources/intentionDescriptions/FillMatchArmsIntention/before.rs.template
@@ -1,3 +1,0 @@
-fn foo(opt: Option<i32>) {
-    <spot>match x { }</spot>
-}
