--- src/main/resources/intentionDescriptions/FillMatchArmsIntention/description.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-This intention fill match arms for enum expression.
-</body>
-</html>
